+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 ### Fixed
 
 - [PostgreSQL Dialect] Fix `string_agg` function to be nullable (#6340 by @griffio)
-
+- [SQLite Dialect] Fix SQLite 3.44 aggregate functions using `GROUP BY` (#6343 by @griffio)
 
 ## [2.4.0-rc1] - 2026-09-01
 [2.4.0-rc1]: https://github.com/sqldelight/sqldelight/releases/tag/2.4.0-rc1